scaler/tests/requirements.txt
grpcio==1.62.1
protobuf==4.25.3
google==3.0.0
vastai==0.2.0
docker==7.0.0
loguru==0.7.2
pytest==8.1.1